backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# ── Validation & Config ───────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
email-validator==2.1.1

# ── Auth & Security ───────────────────────────────────────────────────────
PyJWT==2.8.0
bcrypt==4.1.3
cryptography==42.0.8
python-jose[cryptography]==3.3.0

# ── Cache & Queue ─────────────────────────────────────────────────────────
redis==5.0.4
celery==5.4.0
celery[redis]==5.4.0

# ── HTTP Client ───────────────────────────────────────────────────────────
httpx==0.27.0
tenacity==8.3.0

# ── Object Storage ────────────────────────────────────────────────────────
boto3==1.34.131
aiobotocore==2.13.0

# ── Media Processing ──────────────────────────────────────────────────────
Pillow==10.3.0
aiofiles==23.2.1

# ── AI ────────────────────────────────────────────────────────────────────
openai==1.30.1

# ── Language Detection ────────────────────────────────────────────────────
langdetect==1.0.9

# ── Rate Limiting ─────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Logging ───────────────────────────────────────────────────────────────
structlog==24.1.0

# ── Push Notifications ────────────────────────────────────────────────────
pywebpush==2.0.0

# ── Testing ───────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.0
